Add files via upload
Face Mask Detection dataset:
A face mask detection dataset consists of “with mask” and “without mask” images. I have used the dataset to build a face mask detector with computer vision and deep learning using Python, OpenCV, and TensorFlow/Keras.
The dataset we’ll be using here today was created by PyImageSearch reader Prajna Bhandary.
This dataset consists of 1,376 images belonging to two classes:
• with_mask
: 690 images
• without_mask
: 686 images
Our goal is to train a custom deep learning model to detect whether a person is or is not wearing a mask.
PROJECT CONTENTS:
Dataset - directory contains the data of images of with_mask and without_mask.
Examples – directory contains demo images and the images given in the assignment.
I’ll be reviewing three Python scripts in this assignment:
• train_mask_detector.py: Accepts our input dataset and fine-tunes MobileNetV2 upon it to create our mask_detector.model
• A training history: plot.png containing accuracy/loss curves is also produced
• detect_mask_image.py: Performs face mask detection in static images
• detect_mask_video.py: Using your webcam, this script applies face mask detection to every frame in the stream
Firstly, trained the model for the dataset and got following result:
To accomplish this task, we’ll be fine-tuning the MobileNet V2 architecture, a highly efficient architecture that can be applied to embedded devices with limited computational capacity (ex., Raspberry Pi, Google Coral, NVIDIA Jetson Nano, etc.)
As you can see, we are obtaining ~99% accuracy on our test set.
We can see there are little signs of overfitting, with the validation loss lower than the training loss because the dataset used is of less images. It can be improved by using large data set and training it for more epochs.
Implementing our face mask detector for images with OpenCV
Now that our face mask detector is trained, we have done the following:
1. Load an input image from disk
2. Detect faces in the image
3. Apply our face mask detector to classify the face as either
with_mask
or
without_mask
Results:
As, we can see my model is recognizing only frontline people, not all the people in the image this is limitation of our model. It still needs improvement to detect all the objects in the picture to determine whether they are wearing mask or not.
In order to classify whether or not a person is wearing in mask, we first need to perform face detection — if a face is not found (which is what happened in this image), then the mask detector cannot be applied!
The reason we cannot detect the face in the foreground is because:
1. It’s too obscured by the mask
2. The dataset used to train the face detector did not contain example images of people wearing face masks
Therefore, if a large portion of the face is occluded, my face detector will likely fail to detect the face.
For normal single object static images our model is working absolutely fine:
Implementing our face mask detector for Real time video stream with OpenCV
In python code, detect_and_predict_mask.py
Function accepts three parameters:
• frame: A frame from our stream
• faceNet: The model used to detect where in the image faces are
• maskNet: face mask classifier model